SchemaRDD.cache() now leverages InMemoryColumnarTableScan
